Reorder directories and remove models and processed data from repo
