+++ .gitignore
@@ -1,3 +1,6 @@
+# Project-specific ignores
+outputs/
+
 # IntelliJ stuff
 .idea
