--- http_server/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Test Document</title>
-</head>
-<body>
-    <form action="predict" method="post">
-        <label for="input">StackOverflow question:</label>
-        <input type="text" name="input" id="input" placeholder="What is the answer to the universe?" required><br>
-        <!-- <label for="lastname">Last Name (test)</label>
-        <input type="text" name="lastname" id="lastname" placeholder="Doe" required><br>
-        <label for="url">URL(test)</label>
-        <input type="text" name="url" id="application-url"><br> -->
-        <input type="submit" value="Submit">
-    </form>
-</body>
-</html>
