--- inference/data/processed/tags.txt
@@ -1,100 +1,0 @@
-.net
-ajax
-algorithm
-android
-angularjs
-apache
-arrays
-asp.net
-asp.net-mvc
-c
-c#
-c++
-class
-cocoa-touch
-codeigniter
-css
-csv
-database
-date
-datetime
-django
-dom
-eclipse
-entity-framework
-excel
-facebook
-file
-forms
-function
-generics
-google-maps
-hibernate
-html
-html5
-image
-ios
-iphone
-java
-javascript
-jquery
-json
-jsp
-laravel
-linq
-linux
-list
-loops
-maven
-mongodb
-multithreading
-mysql
-node.js
-numpy
-objective-c
-oop
-opencv
-osx
-pandas
-parsing
-performance
-php
-pointers
-python
-python-2.7
-python-3.x
-qt
-r
-regex
-rest
-ruby
-ruby-on-rails
-ruby-on-rails-3
-selenium
-servlets
-session
-sockets
-sorting
-spring
-spring-mvc
-sql
-sql-server
-string
-swift
-swing
-twitter-bootstrap
-uitableview
-unit-testing
-validation
-vb.net
-visual-studio
-visual-studio-2010
-wcf
-web-services
-windows
-winforms
-wordpress
-wpf
-xaml
-xcode
-xml
