--- inference/src/data/preprocess.py
@@ -1,40 +1,0 @@
-import re
-import nltk
-nltk.download('stopwords')
-from nltk.corpus import stopwords
-from sklearn.feature_extraction.text import TfidfVectorizer
-
-def train_tfidf_vectorizer(X_train):
-    """
-        X_train, X_val, X_test — samples        
-        return TF-IDF vectorized representation of each sample and vocabulary
-    """
-    tfidf_vectorizer = TfidfVectorizer(min_df=5, max_df=0.9, ngram_range=(1,2), token_pattern='(\S+)')
-    X_train = tfidf_vectorizer.fit_transform(X_train)
-    
-    return tfidf_vectorizer
-
-def process_question(text):
-    """
-        text: a string
-        
-        return: modified initial string
-    """
-    REPLACE_BY_SPACE_RE = re.compile('[/(){}\[\]\|@,;]')
-    BAD_SYMBOLS_RE = re.compile('[^0-9a-z #+_]')
-    STOPWORDS = set(stopwords.words('english'))
-    text = text.lower() # lowercase text
-    text = re.sub(REPLACE_BY_SPACE_RE, " ", text) # replace REPLACE_BY_SPACE_RE symbols by space in text
-    text = re.sub(BAD_SYMBOLS_RE, "", text) # delete symbols which are in BAD_SYMBOLS_RE from text
-    text = " ".join([word for word in text.split() if not word in STOPWORDS]) # delete stopwords from text
-    return text
-
-def preprocess_sentences(X_vals, vectorizer):
-    X_vals = [process_question(x) for x in X_vals]
-    X_vals = vectorizer.transform(X_vals)
-    return X_vals
-
-def preprocess_sentence(sentence, vectorizer):
-    sentence = process_question(sentence)
-    sentence = vectorizer.transform([sentence])
-    return sentence
