+++ src/__init__.py
@@ -1,0 +1,3 @@
+
+if __name__ == "__main__":
+    print("Hi")
