+++ src/backend/__init__.py
@@ -1,0 +1,1 @@
+from backend.api import *
